pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "letterboxd-recommender"
version = "0.1.0"
description = "Content-based movie recommendation system built from Letterboxd watch history"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pandas>=2.1",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "beautifulsoup4>=4.12",
    "boto3>=1.34",
    "python-dotenv>=1.0",
    "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = [
    "jupyter",
    "pytest>=8.0",
    "ipykernel",
    "nbformat",
    "matplotlib>=3.8",
    "seaborn>=0.13",
]
ml = [
    "lightfm>=1.17",
    "torch>=2.0",
    "onnx>=1.15",
    "onnxruntime>=1.17",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
